requirements.txt
aiogram==2.27
python-dotenv==1.0.0